# Improve crawling performance of `Notifier#watch`