--- a/lib/rb-inotify/notifier.rb
+++ b/lib/rb-inotify/notifier.rb
@@ -24,10 +24,8 @@ module INotify
   #   # Nothing happens until you run the notifier!
   #   notifier.run
   class Notifier
-    # A list of directories that should never be recursively watched.
-    #
     # * Files in `/dev/fd` sometimes register as directories, but are not enumerable.
-    RECURSIVE_BLACKLIST = %w[/dev/fd]
+    NON_RECURSIVE = "/dev/fd"
 
     # A hash from {Watcher} ids to the instances themselves.
     #
@@ -198,20 +196,17 @@ module INotify
     def watch(path, *flags, &callback)
       return Watcher.new(self, path, *flags, &callback) unless flags.include?(:recursive)
 
-      dir = Dir.new(path)
+      dont_follow = flags.include?(:dont_follow)
 
-      dir.each do |base|
+      Dir.each_child(path) do |base|
         d = File.join(path, base)
-        binary_d = d.respond_to?(:force_encoding) ? d.dup.force_encoding('BINARY') : d
-        next if binary_d =~ /\/\.\.?$/ # Current or parent directory
-        next if RECURSIVE_BLACKLIST.include?(d)
-        next if flags.include?(:dont_follow) && File.symlink?(d)
-        next if !File.directory?(d)
+        next unless File.directory?(d)
+        next if dont_follow && File.symlink?(d)
+        next if NON_RECURSIVE == d
 
         watch(d, *flags, &callback)
       end
 
-      dir.close
 
       rec_flags = [:create, :moved_to]
       return watch(path, *((flags - [:recursive]) | rec_flags)) do |event|