@ioquatix 2024-05-19
I released this in v0.11.1 - if you can test it out that would be awesome. Any further performance improvements are welcome.